[MOV] purchase_alternative_stock*: separate purchase_requisition_stock
*purchase_requisition_stock

With This Commit:
=================================

We have separated the purchase alternative–related functionality from the
purchase_requisition_stock module and reorganised it under a newly created module, purchase_alternative_stock.
This new module depends on both purchase_alternative and purchase_stock,
ensuring proper modular structure and clearer separation of responsibilities.

taskId :- 4737059
