+++ addons/purchase_alternative_stock/__init__.py
@@ -1,3 +1,4 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import test_purchase_requisition_stock
+from . import models
+from . import wizard
