+++ addons/purchase_alternative_stock/__manifest__.py
@@ -1,15 +1,13 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 {
-    'name': 'Purchase Requisition Stock',
+    'name': 'Purchase Alternative Stock',
     'version': '1.2',
     'category': 'Supply Chain/Purchase',
     'sequence': 70,
-    'depends': ['purchase_requisition', 'purchase_stock'],
+    'depends': ['purchase_alternative', 'purchase_stock'],
     'data': [
-        'security/ir.model.access.csv',
-        'data/purchase_requisition_stock_data.xml',
-        'views/purchase_requisition_views.xml',
+        'views/purchase_views.xml',
     ],
     'auto_install': True,
     'author': 'Odoo S.A.',
