+++ addons/purchase_alternative_stock/tests/__init__.py
@@ -1,3 +1,3 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import test_purchase_requisition_stock
+from . import test_purchase_alternative_stock
