+++ addons/purchase_alternative_stock/views/purchase_views.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <odoo>
-    <record id="purchase_order_form_inherit_purchase_requisition_stock" model="ir.ui.view">
-        <field name="name">purchase.order.form.inherit.purchase.requisition.stock</field>
+    <record id="purchase_order_form" model="ir.ui.view">
+        <field name="name">purchase.order.view.form.inherit.purchase.alternative.stock</field>
         <field name="model">purchase.order</field>
-        <field name="inherit_id" ref="purchase_requisition.purchase_order_form_inherit"/>
+        <field name="inherit_id" ref="purchase_alternative.purchase_order_form"/>
         <field name="arch" type="xml">
             <xpath expr="//page[@name='alternative_pos']//field[@name='alternative_po_ids']//field[@name='partner_id']" position="after">
                 <field name='on_time_rate_perc' widget="percentage" invisible="on_time_rate_perc &lt; 0"/>
@@ -11,10 +11,10 @@
         </field>
     </record>
 
-    <record id="purchase_order_line_compare_tree_inherit_purchase_requisition_stock" model="ir.ui.view">
-        <field name="name">purchase.order.line.compare.list.purchase.requisition.stock</field>
+    <record id="purchase_order_line_view_compare_list" model="ir.ui.view">
+        <field name="name">purchase.order.line.view.list.purchase.alternative.stock</field>
         <field name="model">purchase.order.line</field>
-        <field name="inherit_id" ref="purchase_requisition.purchase_order_line_compare_tree"/>
+        <field name="inherit_id" ref="purchase_alternative.purchase_order_line_view_compare_list"/>
         <field name="arch" type="xml">
             <field name="partner_id" position="after">
                 <field name='on_time_rate_perc' widget="percentage" invisible="on_time_rate_perc &lt; 0"/>
