+++ addons/purchase_alternative_stock/wizard/__init__.py
@@ -1,0 +1,1 @@
+from . import purchase_alternative_create
