--- addons/purchase_requisition_stock/wizard/__init__.py
@@ -1,1 +1,0 @@
-from . import purchase_requisition_create_alternative
